pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "energy-orchestrator"
version = "0.1.0"
description = "Self-hosted Python service that minimises electricity cost on a Belgian dynamic-tariff contract by orchestrating SolarEdge / sonnenBatterie / HomeWizard."
requires-python = ">=3.11"
authors = [{ name = "Alex Vinckier", email = "alex.vinckier@kintsunai.com" }]
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.104,<1.0",
    "uvicorn[standard]>=0.24,<1.0",
    "pydantic>=2.5,<3.0",
    "pydantic-settings>=2.1,<3.0",
    "sqlalchemy>=2.0.23,<3.0",
    "aiosqlite>=0.19,<1.0",
    "alembic>=1.13,<2.0",
    "aiohttp>=3.9,<4.0",
    "pymodbus>=3.6,<4.0",
    "websockets>=12.0,<14.0",
    "structlog>=23.2,<25.0",
    "tenacity>=8.2,<9.0",
    "pyyaml>=6.0,<7.0",
    "jinja2>=3.1,<4.0",
    "python-multipart>=0.0.6,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<9.0",
    "pytest-asyncio>=0.21,<1.0",
    "pytest-mock>=3.12,<4.0",
    "pytest-cov>=4.1,<6.0",
    "httpx>=0.25,<1.0",
    "hypothesis>=6.92,<7.0",
    "black>=23.11,<25.0",
    "ruff>=0.1.6,<1.0",
    "mypy>=1.7,<2.0",
    "pre-commit>=3.5,<5.0",
    "types-pyyaml",
]
docs = [
    "mkdocs>=1.5,<2.0",
    "mkdocs-material>=9.4,<10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/energy_orchestrator"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "W",   # pycodestyle
    "F",        # pyflakes
    "I",        # isort
    "B",        # flake8-bugbear
    "UP",       # pyupgrade
    "SIM",      # flake8-simplify
    "C4",       # flake8-comprehensions
    "RET",      # flake8-return
    "ASYNC",    # flake8-async
    "PL",       # pylint
    "RUF",      # ruff-specific
]
ignore = [
    "PLR0913",  # too many arguments
    "PLR2004",  # magic value comparison
    "SIM117",   # combine with statements — combining a server fixture with a client async-with hurts readability
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["PLR", "S101"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
# pymodbus ships type stubs that disagree with runtime in places (e.g.
# ``slave=`` is rejected by the stub but accepted at runtime).  Treat
# the whole package as untyped so we don't fight the stubs.
module = ["pymodbus.*"]
ignore_missing_imports = true
follow_imports = "skip"

[tool.pytest.ini_options]
minversion = "7.4"
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]

[tool.coverage.run]
source = ["src/energy_orchestrator"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
